+++ README.md
@@ -19,16 +19,16 @@ poetry run python scripts/run_embedding_pipeline.py -p "data/test_data/Technical
 Then, you run the RESTful API server with Gunicorn (for production, default worker_class=sync):
 ```bash
 GUNICORN_WORKERS=2 \
-poetry run gunicorn -c gunicorn.conf.py 'scripts.python_helpers.start_api_server:create_app()'
+poetry run gunicorn -c gunicorn.conf.py 'scripts.serving.start_api_server:create_app()'
 ```
 
 Or use the wrapper to pass your usual CLI app arguments along with Gunicorn:
 ```bash
-poetry run -- python -m scripts/python_helpers/run_gunicorn.py \
+poetry run -- python -m scripts/serving/run_gunicorn.py \
   -p data/test_data/TechnicalReport234PDFs \
   -d data/test_data \
   -tm ST -vm CLIP -k 20 -i Memory -- \
-  'gunicorn -c gunicorn.conf.py scripts.python_helpers.start_api_server:create_app()'
+  'gunicorn -c gunicorn.conf.py scripts.serving.start_api_server:create_app()'
 ```
 
 Tuning knobs (Gunicorn env vars supported by `gunicorn.conf.py`):
