+++ compose.yaml
@@ -19,6 +19,6 @@ services:
       # The APP_ARGS variable can be overridden by using an .env file. See https://docs.docker.com/compose/how-tos/environment-variables/set-environment-variables/ for more details.
       - APP_ARGS=--keyword_index_type Lucene --text_model 'Dummy' --visual_model 'Dummy' --backend local --remote_data_directory '/data' --local_data_directory '/data' --host 0.0.0.0
       - GUNICORN_BIND=0.0.0.0:8080
-    command: ["poetry", "run", "gunicorn", "-c", "gunicorn.conf.py", "scripts.python_helpers.start_api_server:create_app()"]
+    command: ["poetry", "run", "gunicorn", "-c", "gunicorn.conf.py", "scripts.serving.start_api_server:create_app()"]
     ipc: "host"
     network_mode: "host"
