+++ govscape/govscape/__init__.py
@@ -16,7 +16,7 @@ from .text_embedding_models import (
     Dummy_TextEmbeddingModel,
     ST_TextEmbeddingModel,
 )
-from .utils import read_txt_file
+from .utils import base_argument_parser, extract_subdomain, read_txt_file, str2bool
 from .visual_embedding_models import (
     CLIP_VisualEmbeddingModel,
     Dummy_VisualEmbeddingModel,
@@ -41,6 +41,9 @@ __all__ = [
     "Server",
     "ServerConfig",
     "WhooshKeywordIndex",
+    "base_argument_parser",
     "build_data_loader",
+    "extract_subdomain",
     "read_txt_file",
+    "str2bool",
 ]
