+++ govscape/govscape/indexing/__init__.py
@@ -1,0 +1,31 @@
+from .keyword import (
+    AbstractKeywordIndex,
+    LanceDBKeywordIndex,
+    LuceneKeywordIndex,
+    SQLiteKeywordIndex,
+    WhooshKeywordIndex,
+)
+from .metadata import (
+    AbstractMetadataIndex,
+    DuckDBMetadataIndex,
+    SQLiteMetadataIndex,
+)
+from .vector import (
+    AbstractVectorIndex,
+    FAISSIndex,
+    LanceDBVectorIndex,
+)
+
+__all__ = [
+    "AbstractKeywordIndex",
+    "AbstractMetadataIndex",
+    "AbstractVectorIndex",
+    "DuckDBMetadataIndex",
+    "FAISSIndex",
+    "LanceDBKeywordIndex",
+    "LanceDBVectorIndex",
+    "LuceneKeywordIndex",
+    "SQLiteKeywordIndex",
+    "SQLiteMetadataIndex",
+    "WhooshKeywordIndex",
+]
