+++ scripts/generate_embeddings.sh
@@ -8,6 +8,6 @@ s3_prefix="s3://bcgl-public-bucket"
 data_dir="test-serving" # 'test-serving', 'dev-serving', or 'prod-serving'
 
 # Run the embeddings pipeline
-poetry run python scripts/python_helpers/run_embedding_pipeline.py --num_pages_to_process 1 \
+poetry run python scripts/pipeline/run_embedding_pipeline.py --num_pages_to_process 1 \
     --batch_size 100 --bucket_name 'bcgl-public-bucket' --pdf_dir 'archive-small/PDFs/' \
-    --data_dir "$data_dir/" --text_model_type 'Dummy' --visual_model_type 'Dummy' --do_text_embedding 1 --do_img_embedding 1 --do_metadata_collection 1
+    --remote_data_dir "$data_dir/" --text_model_type 'Dummy' --visual_model_type 'Dummy' --do_text_embedding 1 --do_img_embedding 1 --do_metadata_collection 1
