+++ scripts/generate_embeddings_distributed.sh
@@ -1,3 +1,3 @@
 
 
-poetry run python scripts/python_helpers/start_embedding_servers.py
+poetry run python scripts/pipeline/start_embedding_servers.py
