+++ scripts/generate_embeddings_docker.sh
@@ -9,7 +9,7 @@ docker run -v ${LOCAL_WORKSPACE_FOLDER:-.}/data:/data --ipc=host --rm -i --gpus 
 
 # Embedding with GPU
 
-poetry run python scripts/python_helpers/run_embedding_pipeline.py --num_pages_to_process 5 \
+poetry run python scripts/pipeline/run_embedding_pipeline.py --num_pages_to_process 5 \
     --batch_size 100 --backend 'local' --local_base_dir '/data/s3_mock' --pdf_dir 'archive/PDFs/' \
     --remote_data_dir "test-serving" --text_model_type 'BGE' --visual_model_type 'CLIP'
