+++ scripts/index_embeddings_img_pg.sh
@@ -10,4 +10,4 @@ embedding_type="/embeddings_img_pg/"
 index_dir="index_img_pg"
 
 # Run the embeddings pipeline
-poetry run python scripts/python_helpers/generate_index_embedding.py --num_pages_to_process 1000000 --bucket_name 'bcgl-public-bucket' --embedding_prefix "$embedding_type" --in_data_dir $data_dir --out_data_dir $data_dir --out_index_prefix $index_dir
+poetry run python scripts/indexing/generate_index_embedding.py --num_pages_to_process 1000000 --bucket_name 'bcgl-public-bucket' --embedding_prefix "$embedding_type" --remote_data_dir $data_dir --out_index_prefix $index_dir
