+++ scripts/index_metadata.sh
@@ -8,4 +8,4 @@ s3_prefix="s3://bcgl-public-bucket"
 data_dir="test-serving"
 
 # Run the embeddings pipeline
-poetry run python scripts/python_helpers/generate_index_metadata.py --bucket_name 'bcgl-public-bucket' --cdx_parquet_key 'archive/2020/CDX/pdf_metadata.parquet' --metadata_prefix $data_dir/metadata --output_prefix $data_dir/index_metadata --output_dir 'data/index_metadata' --num_pages_to_process 50
+poetry run python scripts/indexing/generate_index_metadata.py --bucket_name 'bcgl-public-bucket' --cdx_parquet_key 'archive/2020/CDX/pdf_metadata.parquet' --remote_data_dir $data_dir --num_pages_to_process 50
