+++ scripts/indexing/generate_index_keyword.py
@@ -1,4 +1,3 @@
-import argparse
 import json
 import logging
 import os
@@ -6,6 +5,7 @@ import shutil
 import time
 
 from govscape.data_loader import RemoteDirectoryIterator, build_data_loader
+from govscape.utils import base_argument_parser
 
 import govscape as gs
 
@@ -15,41 +15,16 @@ logging.basicConfig(
     level=logging.INFO,
 )
 
-# ---------------------------------------------------------------------------
-# to run this file: poetry run python s3_ec2_embedding_pipeline.py
-# ---------------------------------------------------------------------------
 
 if __name__ == "__main__":
     # FIELDS TO SET --------------------------------------------------------
-    parser = argparse.ArgumentParser(description="S3 EC2 Embedding Pipeline")
-    parser.add_argument(
-        "--num_pages_to_process",
-        type=int,
-        default=100,
-        help="Number of pages to process from S3",
-    )
-    parser.add_argument(
-        "--batch_size",
-        type=int,
-        default=100000,
-        help="Number of pages to process at a time",
-    )
-    parser.add_argument("--bucket_name", type=str, help="S3 Bucket Name")
-    parser.add_argument("--remote_data_dir", type=str, help="Remote Data Directory")
+    parser = base_argument_parser(description="Generate keyword index")
+    parser.set_defaults(batch_size=100000)
     parser.add_argument(
         "--keyword_index_type",
         type=str,
         default="LanceDB",
         help="Type of keyword index to use: LanceDB, SQLite or Whoosh",
-    )
-    parser.add_argument(
-        "--backend", choices=["s3", "local"], default="s3", help="Data backend to use"
-    )
-    parser.add_argument(
-        "--local_base_dir",
-        type=str,
-        default="data",
-        help="Base directory for local backend",
     )
     args = parser.parse_args()
