--- scripts/python_helpers/__init__.py
@@ -1,1 +1,0 @@
-# Make 'scripts.python_helpers' an explicit Python package for reliable dotted imports
