+++ scripts/server_setup.sh
@@ -13,4 +13,4 @@ s5cmd sync $s3_prefix/index_keyword/* data/prod/index_keyword
 s5cmd sync $s3_prefix/index_metadata/* data/prod/index_metadata
 #s5cmd sync $s3_prefix/metadata/* data/prod/metadata
 
-poetry run python scripts/python_helpers/run_gunicorn.py --data-directory ./data/prod --text_model BGE --visual_model CLIP --keyword_index_type SQLite --vector_index_type Memory
+poetry run python scripts/serving/run_gunicorn.py --data-directory ./data/prod --text_model BGE --visual_model CLIP --keyword_index_type SQLite --vector_index_type Memory
